Support for Selectors API via nwmatcher-noqsa.
Two tiny edits were needed to nwmatcher since our NodeList implementation does
not support array indexing.

It may prove unreasonable to not support array indexing on NodeList, since most
code assumes support despite it not being part of the standard. Yet I do not
believe there is a safe way to support it (i.e., that ignores assignment), so it
may be worth reconsidering how hard we should try to make a strict readonly
implementation--for example, merely documenting that it is unsupported.